Don't check that example configurations compile
Doing this was a problem because after bumping the version, we would run
tests, and we would try and compile examples, which at that point would
include a non-existing version of the package.

In practice, since we generate examples from preview, I can't think of a
case where examples would not work when preview does.