Running local version of optimize
I just tried including `elm-optimize-level-2` inside of `elm-review`.
The results look something like this (analyzing a 160k LoC project):
Before optimization, the running of the Elm code takes 5.23s-5.40s
After optimization, the running of the Elm code takes 4.75s-4.9s
So it's a bit less than I had hoped for :stuck_out_tongue: and also, the 
optimization takes 2 seconds, so the initial run is way slower (~1.5 
seconds for this project, but ~2 seconds on small project where the 
gains are much less) on the initial run (the following run we don't 
re-compile).

Note that:
- I'm compiling the JS THEN recompiling, as I need a bit more leeway to 
the node-elm-compiler settings
- I am spawning a new process, as I can't use `elm-optimize-level-2` as 
a Node.js library with the current implementation

So all-in-all, I am not yet convinced, since I care about compilation 
times.

For comparison, the analysis of small packages takes like 275ms-290ms 
without optimization, and ~240ms with, with still somewhere between 
1.6s-2s of compilation time.