Change the app hash for remote templates to include the path
This prevents a crash when using two different template configuration
at different folder paths (/preview1 and /preview2) and re-using the
result cache.

The crash happens because we would be using the result tied to the built
application of a different compilation, meaning some fields would be
mismatched and result in "field is undefined"-like errors.

We solve this by hvaing both the path to the folder and the commit to
the path of the result cache, removing the confusion.