chore: fix ls-engines
Ok, so this one is complicated.
SemVer wise, we should be able to support glob 10 & glob 11.
The only breaking change is the engine.
Therefore, the "correct" specifier is indeed `^10.2.6 || ^11.0.0`.
We support both, if you need an older engine,
you should use an older version.
In fact, npm is supposed to do this automatically.
However, up until `npm-pick-manifest@9.1.0` (included in `npm@10.8.2`),
there was a longstanding bug where it didn't!
The code was there,
but the logic didn't work right and after Isaacs moved on to vlt,
his PR fixing this got stalled.
After the `glob` fiasco, he rebased the PR and it's now fixed,
but the change wasn't backported.
Because we support Node 14, we support `npm@6.14.18`,
so we can't take advantage of this.
Instead, we just have to tolerate a slightly old version of glob.
Oh well.

I also removed "enginesStrict" again
because our engine declaration is now higher than technically needed
because `glob` only declares support for the latest versions of Node branches.