feat: port result from elm/core
Uhhhh......
Effect is heavy, and TS didn't like the way message passing was done...
And, so I uh... wrote a result type.
And then I made it generic, so voilà!

The only issue is that TS can't type pipes,
so I'd need to make a fluent interface or write a ton of overloads.
I opted to just force normal functions.