refactor: make result even more elm-y
I started with uppercasing the `R`.
Then, I replace  `Err` with `Failure` so it doesn't get mixed up with native JS errors.
I did the same with `Ok` and `Success`, so that the constructors line up.
I tried uppercasing the constructors as well, to make it more elmish.*
However, Eslint triggers `new-cap`, so I skipped that one and kept the Task terminology.
Also, because objects are really inconvenient,
I moved it to a folder and used evil barrel exports.

* I tried to come up with a Fable pun, but I couldn't think of one.