refactor: use zx
zx is a much simpler way to run commands.
It's a much higher level library than child_process,
so the code is much simpler to read.
Length-wise, it's about the same, but slightly longer,
primarily because Prettier wraps lists but not strings.
I also converted the remaining bits of sync io to async io,
as async io is only faster the more you use it.
It also encourages function coloring,
which makes it more explicit where there's io.
(Yes, I know function coloring is generally a bad thing.)